put vehicle burn under token endpoint; web2 delete
